refactor tree search into own file
