+++ app/assets/javascripts/oxalis/view/right-menu/trees_tab_view.js
@@ -3,10 +3,9 @@
  * @flow
  */
 
-import _ from "lodash";
 import * as React from "react";
 import { connect } from "react-redux";
-import { Button, Dropdown, Input, Menu, Icon, Spin, Modal, AutoComplete, Popover } from "antd";
+import { Button, Dropdown, Input, Menu, Icon, Spin, Modal } from "antd";
 import TreeHierarchyView from "oxalis/view/right-menu/tree_hierarchy_view";
 import InputComponent from "oxalis/view/components/input_component";
 import ButtonComponent from "oxalis/view/components/button_component";
@@ -33,7 +32,7 @@ import Toast from "libs/toast";
 import { getBuildInfo } from "admin/admin_rest_api";
 import type { Dispatch } from "redux";
 import type { OxalisState, SkeletonTracingType, UserConfigurationType } from "oxalis/store";
-const Option = AutoComplete.Option;
+import TreeSearchPopover from "./tree_search_popover";
 
 const ButtonGroup = Button.Group;
 const InputGroup = Input.Group;
@@ -56,67 +55,6 @@ type State = {
   isUploading: boolean,
   isDownloading: boolean,
 };
-
-const maxSearchResults = 10;
-class TreeSearchPopover extends React.PureComponent<*, *> {
-  autoComplete: *;
-
-  constructor() {
-    super();
-    this.state = {
-      isVisible: false,
-      searchQuery: "",
-    };
-  }
-
-  render() {
-    const filterOption = (inputValue, option) => option.props.text.indexOf(inputValue) > -1;
-    const filterTree = tree => tree.name.indexOf(this.state.searchQuery) > -1;
-
-    return (
-      <Popover
-        title="Search Trees"
-        trigger="click"
-        placement="rightTop"
-        visible={this.state.isVisible}
-        mouseLeaveDelay={1}
-        onVisibleChange={isVisible => this.setState({ isVisible })}
-        content={
-          // Only render autocomplete when the popover is visible
-          // This ensures that the component is completely re-mounted when
-          // the popover is opened. Thus, autoFocus works and unnecessary
-          // computations are avoided.
-          this.state.isVisible && (
-            <AutoComplete
-              autoFocus
-              onSearch={searchQuery => this.setState({ searchQuery })}
-              dataSource={_.values(this.props.trees)
-                .filter(filterTree)
-                .slice(0, maxSearchResults)
-                .map(tree => (
-                  <Option key={tree.treeId} text={tree.name}>
-                    {tree.name}
-                  </Option>
-                ))}
-              style={{ width: "500px" }}
-              ref={autoComplete => (this.autoComplete = autoComplete)}
-              onSelect={(value, option) => {
-                console.log("onSelect", value, option);
-                this.props.onSetActiveTree(option.key);
-                this.setState({ isVisible: false });
-                if (this.autoComplete) this.autoComplete.blur();
-              }}
-              placeholder="Input here to search trees"
-              filterOption={filterOption}
-            />
-          )
-        }
-      >
-        {this.props.children}
-      </Popover>
-    );
-  }
-}
 
 class TreesTabView extends React.PureComponent<Props, State> {
   state = {
@@ -266,8 +204,9 @@ class TreesTabView extends React.PureComponent<Props, State> {
         </Modal>
         <ButtonGroup>
           <TreeSearchPopover
-            onSetActiveTree={this.props.onSetActiveTree}
+            onSelect={this.props.onSetActiveTree}
             trees={this.props.skeletonTracing.trees}
+            maxSearchResults={10}
           >
             <ButtonComponent>
               <Icon type="search" />
